board/hyperdebug: Support slow I2C clock speeds
When using HyperDebug on some development boards that lack proper I2C
pullup, we may want to set the clock speed slower than 100kbps, which is
currently the lowest setting.

This CL makes use of I2C clock prescaling 1-16x using the high 4 bits of
the TIMINGR register to be able to scale the pre-calculated high/low
timing values for 100kbps down to 50kbps, ..., 6.26kbps.

In order to achieve the above, the logic of chip_i2c_set_freq() has been
copied into board/hyperdebug/i2c.c, as this functionality is probably
not something we want to support in the general EC codebase.  (It would
require replacing `enum i2c_freq` with an integer, which would be
non-trivial.)

BUG=b:309485648
TEST=observe I2C signal on logic analyzer

Change-Id: I86c58efd31acaaace6a04e15b5962079233529b5