TCPMv2: Handle SOP' Discover ID response in DPM
Record the response (or non-response) to SOP' Discover Identity in the
PE by calling into the DPM, which in turn records the response properly.
This is a first step in moving discovery policy into the DPM, including
the timing of initiating discovery messages. The bulk of this commit
creates the discovery state structure in the DPM.

This commit does not change the observable discovery policy of the TCPM,
but it is a step towards making it easier to do so.

BUG=b:272827504
TEST=Attach to dock; observe initial PD traffic, including discovery

Change-Id: I76b140cc1ff9382898b5e4bee00ca878d1fc83fa