craask: Add touch panel power sequence control
To ensure that EC can quickly resolve touchscreen power sequence issues
such as b/301372546, the touchscreen enable pin is routed to the EC.
The EC will add a delay time when enabling the touchscreen.
In the future, adjusting this delay may become an option for resolving
touchscreen power sequence issues.

BUG=b:313494799
TEST=./twister -T zephyr/test/nissa

Change-Id: I36778142b6dc2420a9599b83d19227424387602c