fan: always initialize speed reports to NOT_PRESENT
If the system fan count is set to zero during init before pwm_fan_init
runs at HOOK_PRIO_DEFAULT, the init function would skip setting
memmap-reported speeds to reflect that the fan(s) are absent. This
manifests as incorrect quantities of fans reported by ectool on nissa
and fizz boards that are fanless where the presence or absence of a fan
is detected at runtime.

Fix this by always running full fan init, even if there are no fans.

BUG=b:308363404
TEST=Fanless pujjoteen now reports 0 fans rather than 4;
     ./twister -T zephyr/test/drivers -s drivers.fan;
     Added test case fails without init changes

Change-Id: Id233e05f887cf201e8481373b8b240b83d93404a