Add a host command to print info about AP-firmware state
Provide a simple host command that prints a 32-bit hex value on the
console with an 'AP_FW' prefix. This can be used by AP firmware to tell
the EC things about its current state.

FAFT can then monitor the EC console to see the current state of the AP
firmware.

The EC does not understand the value. It just prints it in hex format.
This avoids the EC needing to share header files or other code with AP
firmware.

BUG=b:300525571
TEST=./twister -p native_posix -s drivers/drivers.host_cmd

Change-Id: Ic42d667fd15be56772941d75218bf736fa83fd75