fpsensor: Annotate function provided from static library
It is very confusing to understand what functions are provided by the
private libraries and which are abstractions that we have created.

We will now mark functions that have no source and are provided by
static libraries as __staticlib. Functions that are only intended to
serve as a hook for the staticlib will be marked as __staticlib_hook.

The definition of these macros is less important than the annotation
itself. Feel free to make these blank in the future.

BUG=b:295249061
TEST=make buildall
TEST=./util/compare_build.sh --board=fp --noprivate --ref1=cros/main \
       --ref2=HEAD
TEST=./util/compare_build.sh --board=fp --private --ref1=cros/main \
       --ref2=HEAD

Change-Id: I905a5b71d053439f261dcdaddaed2f818e3635c3