zephyr: driver: cros_shi: npcx4: add enhanced buffer mode
The original SHI module only has one output FIFO buffer. It costs a lot
when the driver has to send/change the protocol control code because it
must fill out all 128 bytes of output FIFO. In npcx4, we introduce
another output buffer in 1-byte depth. These two buffers can switch back
and forth during the transaction. We can use the single-byte buffer
to send the control code and the 128-byte FIFO to send the data payload.
It helps improve the SHI driver's efficiency.

This change depends on upstream change
    https://github.com/zephyrproject-rtos/zephyr/pull/65557/

BRANCH=none
BUG=none
TEST=use the host emulator to send host command.
LOW_COVERAGE_REASON=this is the chip-specific hardware driver and
cannot be tested via emulator.

Change-Id: I562419d2906f4b87121ccbfc4589f6adc3409565
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>