zephyr: add explicit kso2 pin mapping for ite based devices
The ITE raw keyboard scan driver can handle the push-pull inverted KSO02
signal directly and does not use a GPIO to use that.

Drop the note about the signal handled by a GPIO, drop the GPIO
definitions for it and add a pinctrl node to map it to the keyboard
scanning device explicitly.

BUG=b:274443029
TEST=cq dry run

Change-Id: Ia0dc206fe7446a7ebe5c1569e911edc464ec68af