zephyr: fpmcu: Enable DMA for console UART for bloonchipper
https://github.com/zephyrproject-rtos/zephyr/pull/63967 adds support for
UART asynchronous API (if enabled) in shell.

In bloonchipper, we use USART1 to communicate with AP and USART2 for
console. We are using DMA to communicate efficiently with AP, so we
must have UART asynchronous API enabled. Previously, this was not
a problem because UART backend for shell was not using it, but #63967
broke shell due to missing DMA configuration.

The DMA engine is shared with SPI connected to the fingerprint sensor.
Priority for USART2 DMA requests was set to low priority.

BUG=b:299924239
TEST=Build firmware for bloonchipper using `zmake build bloonchipper`.
     Flash firmware. Make sure shell is working properly.

Change-Id: I1c71b0e8e9a8b28eec78ade4af4bacc4124751d2