Karis: Add touch panel power sequence control
The touchscreen enable pin is routed to the EC.
The EC enable the touchscreen while the panel backlight is
turned on by SOC. The EC disable the touchscreen while
the lid is closed or the SOC turn of the panel backlight.
The EC will add a delay time when enabling the touchscreen.
In the future, adjusting this delay may become an option for
resolving touchscreen power sequence issues.

BUG=b:314245238
TEST=./twister -T zephyr/test/rex
TEST=set FW_CONFIG bit14=1 on re-work board.
     1. check if the soc_3v3_edp_bl_en interrupt triggered
     2. check ec_touch_en, soc_3v3_edp_bl_en and lid status.
        soc_3v3_edp_bl_en=1, lid=open -> ec_touch_en=1
        soc_3v3_edp_bl_en=0, lid=open -> ec_touch_en=0
        soc_3v3_edp_bl_en=1, lid=close-> ec_touch_en=0
        soc_3v3_edp_bl_en=0, lid=close-> ec_touch_en=0

Change-Id: I8ceb17ebc9fc102aaea3d4634171fdca8f469cca