board/hyperdebug: Fix detail in CMSIS-DAP JTAG protocol
Existing implementation had a bug in the way that JTAG data from the
device under test was packed into bytes of the USB response.

In case of several "sequences" each having a clock count that may not be
divisible by 8, then the proper protocol is to pad the data from each
sequence to a full byte, before sending the data from the next sequence.

BUG=none
TEST=make BOARD=hyperdebug
TEST=manually verified using OpenOCD  (by awill@opentitan.org)

Change-Id: Ib4eb42cccae15ff6284fd1d1c071f11d7ee8931a