aprv: return the saved hash if the contents are valid
The hardware may not support verification, but there may be a valid hash
saved in cr50. Let the user check the saved hash even if the hardware
doesn't support verification. The user can use the get status vendor
command to see if verification is supported. The get hash command should
return the hash.

BUG=b:312070739
TEST=get hash on a device that doesn't supportverification

Change-Id: I3101da5e5629df729489035a58f5bc1812aeec57