ap_pwrseq: enable driver model by default
The driver variant of ap_pwrseq is meant to be the long-term solution
for power sequencing by improving on a number of API aspects. However,
it's unlikely that existing platforms will migrate to it due to the risk
involved in changing power sequencing implementations. To encourage new
platforms to use the correct implementation, change the driver model to
be the default and explicitly disable it for existing platforms that use
ap_pwrseq.

BUG=b:217952699
TEST=zmake compare-builds -a

Change-Id: I76208bae9bc1bc0f9ab99db5314fb64145811871