ectool: Allow commands to parse their own options
Currently, if a non-option is encountered, ectool prints 'invalid
option' and exits. For example, the '-c' in 'ectool bcfg get -c' is
processed by getopt_long in the main() and produces an error.

This CL makes getopt_long in the main() stop parsing options at the
first non-option, which should be a command name. It will allow commands
to parse their own options.

There is no functionality change.

BUG=b:285378416, b:314827947
TEST=Run ectool bcfg get -c

Change-Id: I0a3612d3224b4ab33ac2d9ae5042c79315d6a3ee