test/run_device_tests.py: Pass BoardConfig instead of board string
In the special case of the "nocturne_fp" and "nami_fp" tests, we build
for nocturne_fp/nami_fp, but the BoardConfig is for dartmonkey.

BOARD_CONFIGS[build_board] in run_test will fail because
nocturne_fp/nami_fp do not exist in BOARD_CONFIGS. Instead, we can just
pass through the original board_config to run_test.

Now that we're passing the BoardConfig to run_test, we can get the
reboot timeout from that rather than passing it as a separate argument.

BUG=b:314810166
TEST=./test/run_device_tests.py --board dartmonkey \
       -t unaligned_access_nocturne_fp_v2.2.64

Change-Id: Ibb00867d8019b1537bef4bbe954c3e38eff5a82d