Cutoff: Don’t cutoff battery if AC is disconnected temporarily
Currently the battery cutoff manager is using
HOOK_POWER_SUPPLY_CHANGE to monitor the active charging port.
However, there is a chance if AC is disconnected temporarily such
as hard reset the cutoff manager will false triggering the battery
cutoff. To avoid this, we should cancel the cutoff event if AC is
disconnected temporarily.

BUG=b:314872582
TEST=On Quandiso.
1. Cutoff by Refresh+Unplug AC.
2. No cutoff by holding Refresh down until boot into depthcharge.

Change-Id: Ia07c2a9fb8b30e298ed16ee70029908e330acbef