third_party/rules.mk: Disable filesystem support in googletest
EC doesn't have a filesystem. The googletest code for filesystem support
increases code size, so disable it to free up flash space.

With this change googletest prints the following message at the start of
the test:

[ ERROR ] gtest.cc:5650:: ERROR: alternative output formats require
                          GTEST_HAS_FILE_SYSTEM to be enabled

but it has no effect on the tests and looks like a bug that has been
fixed in newer versions of googletest:

https://github.com/google/googletest/issues/4299
https://github.com/google/googletest/commit/2acd53824402affcf473b2355d7323047e660e86

BUG=b:254530679
TEST=make BOARD=bloonchipper test-fpsensor_hw -j
TEST=make BOARD=dartmonkey test-fpsensor_hw -j
TEST=./test/run_device_tests.py --board bloonchipper -t fpsensor_hw
TEST=./test/run_device_tests.py --board dartmonkey -t fpsensor_hw

Change-Id: Ieec4786f6073d04bb3424b36aa6d811a8c572465