bindings: fingerprint: Move GPIO properties to devices
Properties 'irq-gpios' and 'reset-gpios' are associated with the device
and shouldn't be considered as a general fingerprint sensor properties.

For example, it doesn't make sense to define GPIO properties for
fingerprint sensor simulator because it's virtual device used in unit
tests.

Defining GPIO properties in device bindings is common pattern in
Zephyr.

BUG=b:299924239
TEST=zmake build bloonchipper

Change-Id: Ib45ceade855c5e8ca26d53b34a285bc1fc51bd97