zephyr: Disable input prompt for FINGERPRINT_SENSOR_IMAGE_{SIZE,OFFSET}
The FINGERPRINT_SENSOR_IMAGE_SIZE and FINGERPRINT_SENSOR_IMAGE_OFFSET
Kconfig options were introduced as configurable by the user.

Actually these options are used to override image size and highly depend
on the sensor model, so changing the defaults can make the driver to
work improperly. If the correct image size can be obtained from DTS, we
should avoid providing a default values here.

If the option with the 'int' type is user configurable, Kconfig will
always emit appropriate CONFIG_ define even if the value is not set and
there is no default. This causes problems with determining if we should
use CONFIG_FINGERPRINT_SENSOR_IMAGE_SIZE or obtain image size from DTS.
This problem doesn't occur if the option is not user configurable.

BUG=b:299924239
TEST=zmake build bloonchipper

Change-Id: Ibe158a329540c0cf31cbb5f5c5adea216e80e6d8