TCPMv2: Disable USB mux for USB PD partners which do not support USB
Updates the USB mux setting to disable the mux for USB PD partners which
don't set the USB Communication Capable bit. This also adds a condition
to prevent setting the mux when in a TC_CT_* state. Previously this
would be seen as a new PD connection and lead to the mux getting cleared
when connecting a charge through device.

BUG=b:309881356, b:309706714
TEST="./twister -p native_posix -p unit_testing", "make runhosttests",
"make try_build_boards", flashed brya and confirmed the USB mux was 0
for USB PD chargers and 1 for non USB PD devices.

Change-Id: I32e5687b5be56301490d880208bc91e13e882686