cec: Drop support for cec_message MKBP event.
CL:4659321 added support for receiving CEC messages on multiple ports.
Before that change, messages were transferred to the AP using the
cec_message MKBP event. Now we use EC_MKBP_CEC_HAVE_DATA to notify the
AP that data is ready, and the AP reads it using a read command.

To make it easier to roll out the changes, that CL continued to use
cec_message on devices with only one port, so that it doesn't matter
which order the EC and kernel changes land.

Multi-port support in the kernel landed on all branches 5.4+ a couple of
months ago (CL:4951179 and cherry-picks), so it should be safe to drop
cec_message support in the EC now.

The kernel and ectool will continue to support both methods because it's
hard to guarantee that EC multi-port support has propagated to all
devices.

BUG=b:296314584
TEST=CEC works on kinox

Change-Id: I971b8f003b7fec4b0ef11c1189c01ef2a23d8dc7