keyboard: add a devicetree entry for vivaldi keyboard
Add a devicetree based Vivaldi keyboard binding an driver.

This supports implementing a custom Vivaldi configuration with no board
code if only one config is needed, and with a single callback function
to select the active config if there are multiple configurations and the
right one has to be selected in runtime.

BUG=b:274443029
TEST=cq dry run

Change-Id: I7083cba3274a1b24ec00081f284e96fb004d16af