board/hyperdebug: Fix bug in handling of capabilties query
HyperDebug implementes some Google extensions to the CMSIS-DAP protocol,
among those a method to query which specific Google extensions are
supported.  This query method however was not initially used by
`opentitantool`, and it seems a bug has crept in, causing it to remove
more bytes from the incoming stream than the two being sent.

BUG=none
TEST=Run modified `opentitantool`, observing capabilities bitmask

Change-Id: I2671b71e309c8a639b8c2436bb83bee31078377f