test: Add clear_fpscr function and call it for all tests
Introduce the clear_fpscr function to clear the fpscr register. Call
this function before running the tests for the STM32 targets. The
fpu_irq() handler clears the FPSCR register, making the call to
clear_fpscr unnecessary. However, we need to run these tests on chips
that do not have FPU interrupt. As such, this function becomes
essential.

BUG=b:298697667
TEST=run_device_tests.py --board dartmonkey --flasher=servo_micro
TEST=run_device_tests.py --board bloonchipper --flasher=servo_micro

Change-Id: I61bd9036254a0b81fe494e15b154766beaf88af5