SBS: Disable alarm warning
The smart battery specification specifies "the ALARM_MODE bit is
automatically cleared by the Smart Battery electronics every 60 seconds
so that ... A SMBus Host which does not want the Smart Battery to be a
master on the SMBus must therefore continually set this bit at least
once per 45 seconds...".

With this CL, the EC will set the ALARM_MODE bit in the mode register
every 45 seconds to keep alarm warning disabled.

BUG=b:188662830
TEST=Vell

Change-Id: Ia97dccac97e9409fb1c67c5e821e03f0eee12aaf