Charger: Reinitialize battery when responsiveness changes
Currently, the battery is (re)initialized only when the battery's
physical presence changes. This causes the charge task to fail to
configure the battery when the battery doesn't immediately responsive
after it becomes physically present.

This CL makes the charge task call process_battery_present_change when
BATT_FLAG_RESPONSIVE changes as well.

BUG=b:315883344
TEST=Yaviks + SMP. Cutoff at-shutdown. Plug AC. Verify by bcfg command.

Change-Id: I36a693c1a13a0df61a8ef17b15a7825a71be1746