nissa/yavilla: fix races in power LED control
The previous handling of power LED state changes was prone to races when
running on Zephyr using the new ap_pwrseq power sequencing
implementation because ap_pwrseq runs HOOK_CHIPSET_ hooks from the power
sequencing thread, not from a workqueue. If the AP power state changed
and preempted the handler for a power LED deferred funcction, the
attempt(s) to cancel the preempted handler would fail (because it's in
the "running" state) and the LED could be left in an incorrect state.

This change fixes those races by using a Zephyr timer for suspend and
shutdown delays to ensure it can always be cancelled (because timer
handlers run in interrupt context, they cannot be preempted by hooks),
and all mode changes are performed atomically by requesting the main
power LED tick function change state rather than changing the state
directly which might race with active work to pulse the LED.

BUG=b:305126895
TEST=zmake build yavilla;
     visual inspection of LED on hardware seems to show normal behavior
LOW_COVERAGE_REASON=Bugfix to device-specific code that never had any tests

Change-Id: I75526989a5e042f94bfdcb34906116a7ea9fc4b1