Battery/config: Fix device name comparison
Currently, device names are strictly tested. That is, the two strings
have to match including the lengths. This CL makes the comparison
partial matching (or prefix matching).

Battery's device name should be able to contain extra trailing
characters. That is, if the battery says "xyz 123", a config whose
device name is "xyz" should match.

This is consistent with the legacy mode (configs embedded in FW).

BUG=b:283131235
TEST=make run-battery_config

Change-Id: I19c54a83440b3dd2a2ef26d10092ca10808a9e69