zephyr: cros_flash: Define the size to reload the watchdog timer
We reload the watchdog timer to prevent any reset during a long erase.
The reload size is not necessary to be the same size as the erase size.
Define a separate constant for the reload size. So changing the erase
size won't affect the watchdog reload.

BUG=b:315899470
TEST=zmake compare-builds --ref1 HEAD~ --ref2 HEAD rex
LOW_COVERAGE_REASON=Define constant on code w/o coverage, b/292126499

Change-Id: I8897cecd0e9e56c5c25f526b67cb59dddb5c9067