ec: Fix builtin stdlib on host BOARD and tests
If the board is host, disable USE_BUILTIN_STDLIB by default.

Change the stdlib test to #include the .c file to be
tested, even though that is gross, but it gets around the symbols being
defined both in the EC code and also the stdlib.

Fix the rule for including shmalloc.c in common/build.mk

BUG=b:237823627
TEST=make -j$(nproc) clobber buildall

Change-Id: Id4d9be6ed296a1d4668982f149884e0c274d0a69