fpsensor: Add support for ELAN 80SG sensor
The SRAM available for the Nuvoton M4 MCU is less than that of
the STM32 M4/M7 MCU. With less SRAM, the number of fingerprint
templates and the overall performance will need to be decreased.

The new config "CONFIG_FP_SENSOR_ELAN80SG" can be used to enable
this support with the proper parameters.

BUG=b:306455578
TEST=# Change board/helipilot/board.h to have the following:
     # #define CONFIG_FP_SENSOR_ELAN80SG
     make BOARD=helipilot

Cq-Depend: chrome-internal:6837675
Change-Id: I5e8e06c57daa95d678f4e4a689d42c43696c5a42