pdc: Make API safer for concurrency, add completion sempahore
Check if there is a pending command loaded before setting up a new one
to avoid a race condition during the period of time before the state
machine begins processing it.

Add a sempahore that is given when a command finishes processing and a
helper function to take / wait on this semaphore. This allows clients of
the driver to interact with it synchronously.

BUG=b:318699541
TEST=zmake build brya_pdc

Change-Id: I5c19a5045cba64a5f664869cddd11047546c9da7