zephyr: bloonchipper: disable RTC access in runtime
Disable APB clock for RTC in runtime to save power, since it is not used
anyway. Reenable the RTC APB clock based on the sleep pin, because it is
needed to wake up the chip from the stop mode. The APB is used to access
the RTC registers, the RTC counter is running without any break.

It saves ~0,5mW in runtime.

BUG=b:311568657
TEST=Measure the power consumption with command "dut-control -t 60
ppvar_mcu_mw" and make sure it drops ~0,5mW in runtime.

Change-Id: Ib3bf730eb1c52bc1c70f01933f33489845156013