board/hyperdebug: Avoid panics on externally jammed I2C bus
While HyperDebug acts as I2C device, it may not have data
ready in case the I2C host asks for it, and will stretch the
I2C clock waiting to receive data via USB.  We use a timeout
feature of the I2C IP block to put a bound on the time that
HyperDebug can stay in that state, after which it releases the
SCL line, and assumes that the I2C host has given up receiving
data anyway.

It turns out that prior to this CL, the timeout feature used
would also trigger if the clock line was pulled low by the I2C
host for an extended period (or by pullups being unpowered).
This confused our state machine (resulting in panic), as it
only expected the timeout interrupt while an I2C read
operation was in progress.

Fortunately, the STM32L5 microcontroller has a second kind of
timeout setting, that only counts the time when HyperDebug
itself keeps the clock line low.  This CL enables that
timeout, instead of the broader one.

BUG=none
TEST=make tast TAST_EXPR=gscdevboard.GSCI2CBridge.cap_always  (in ti50/common)

Change-Id: Ic4a95173928e59e5ab08435f77a1660c45e80dcd