zephyr: cros_flash: Use the smallest erase size for alignment check
In the NPCX cros_flash implementation, the CONFIG_FLASH_ERASE_SIZE is
used for alignment check only. The actual erase is done in the Zephyr
upstream driver, which doesn't use this CONFIG value. Instead, it
dynamically picks the best erase size from 4KB(SE) or 64KB(BE).

So use the smallest one for alignment check.

BUG=b:315899470
TEST=Modify a rex project to use the `npcx9m3f` board, put the
CBI portion to the end of the RO, i.e. offset 0x3f000, build its
image, and verify the CBI portion.

Change-Id: I8fe067e9726a6200dc2d08d1d23dc24f25d138ba