ps8743: Limit USB mode update to USB2 connection on resume
This is a follow-up of CL:5172430.

In ps8743_suspend(), PS8743_REG_MODE field gets updated only for USB2
connections. However in ps8743_resume(), the field gets updated for
all states as long as there's a device connected.

This may lead to unbalanced PS8743_MODE_USB_ENABLE bit write to the mux
on every resume despite the driver didn't disable USB mode on suspend
for USB3 connection.

We haven't seen any report saying this specific behavior causes issues,
but we just saw an example in CL:5172430 that the excess USB mux update
event cause suspend to fail with external DP display attached.

Limit the branch condition to USB2 connections only. This better aligns
with the logic on the suspend side and skip unnecessary reg field update
on resume for USB3 connections.

BUG=b:316829805
TEST=zmake build ciri --clobber

Change-Id: I17d8bf94ec5f1285922deddd0a7fd5b382f5a056