board/hyperdebug: Fixes to code path for buffer overrun
It appears that in past refactorings, some bugs had been introduced in
the uncommon path handling the case when opentitantool does not retrieve
GPIO monitoring events in time, and the internal buffer overruns.

We want to handle that case by stopping monitoring, and reporting the
overrun next time opentitantool attempts to retrieve events.

BUG=b:266832220
TEST=make tast TAST_EXPR=gscdevboard.CCDCapabilitiesFlashEC.cap_always
	  (with local modifications to capture ITE waveform)

Change-Id: Ie5e82d6e47c4349239b7b268e097ea9b5c2693ba