test: Add clear_fpscr function and call it for all tests
Introduce the clear_fpscr function to clear the fpscr register. Call this
function before running the tests for the STM32 targets. The fpu_irq()
handler clears the FPSCR register, making the call to clear_fpscr
unnecessary. However, we need to run these tests on chips that do not
have FPU interrupt. As such, this function becomes essential.

BUG=b:298697667
TEST=run_device_tests.py -b bloonchipper -f servo_micro
TEST=run_device_tests.py -b dartmonkey -f servo_micro

Change-Id: Id04e356333e30dec7e4a8a9149aa13f4d85ab219