npcx9m7f: Move FMAP to 0x4fc00
The FMAP data structure can be placed anywhere in the flash image
because tools like cbfstool and dump_fmap do a full search of the flash
image for the "__FMAP__" marker to locate it. However, this search is
very expensive because the entire flash image must be read before these
tools can be used.

As an optimization, tools like flash_ec use the FMAP offset of a new EC
image as a hint to locate the FMAP in flash. If this hint location
contains the FMAP, it is used. Else, the entire flash is read and
dump_fmap is used to locate the FMAP in the extracted image.

binman doesn't have a way to place an object at the end of a section, so
we reserve 1 KB bytes of flash for the FMAP at offset 0x4fc00 which we
know to be at the end of the ec-ro section on the npcx9m7. The FMAP can
vary in size and is currently 560 bytes on rex.

Here's the resulting FMAP:

  [rex] $ dump_fmap -F build/zephyr/rex/output/ec.bin
  0x00000000:0x0007ffff WP_RO
  0x00000000:0x0004fe2f EC_RO
  0x00000000:0x000414df RO_FW
  0x000414e0:0x000414ff RO_FRID
  0x0004fc00:0x0004fe2f FMAP
  0x00050000:0x00050fff CBI
  0x00080000:0x000cffff EC_RW
  0x00080000:0x000c175f RW_FW
  0x00080000:0x000c175f ZEPHYR_RW
  0x00080144:0x00080163 RW_FWID
  [rex] $

BUG=b:317306098
TEST='zmake build -a' passes
	'./twister -v -T zephyr/test' passes
	booted on rex

Change-Id: Ia08a271f011d047035fa1e8c037b4b4c0c82b209