binman: Swap fmap and ro-frid sections
The FMAP should be at a stable offset, else tools need to search the
entire flash image to find it, and that can be non-deterministic because
they basically scan for the "__FMAP__" signature.

By swapping the order, each chip can pin FMAP to the end of its ec-ro
section which makes the offset stable. We also need to set the pad-byte
so that any gaps created by moving the FMAP are padded as expected -
i.e. the same pad-byte as the parent region.

Here's the resulting FMAP:

  [rex] $ dump_fmap -F build/zephyr/rex/output/ec.bin
  0x00000000:0x0007ffff WP_RO
  0x00000000:0x0004172f EC_RO
  0x00000000:0x000414df RO_FW
  0x000414e0:0x000414ff RO_FRID
  0x00041500:0x0004172f FMAP
  0x00050000:0x00050fff CBI
  0x00080000:0x000cffff EC_RW
  0x00080000:0x000c175f RW_FW
  0x00080000:0x000c175f ZEPHYR_RW
  0x00080144:0x00080163 RW_FWID
  [rex] $

BUG=b:317306098
TEST='zmake build -a' passes
	'./twister -v -T zephyr/test' passes
	booted on rex

Change-Id: I9efb77fc3e7aedce011b100b3624c8b7176c2ba5