boot_keys: add a new input based boot_keys driver
Add a new boot_keys driver that reimplements the
CONFIG_KEYBOARD_BOOT_KEYS functionalities currently provided in
keyboard_scan.c. This implementation is not intertwinted with the
keyboard driver itself and is enabled and configured using a devicetree
node, reducing the amount of Kconfig options and the need of a
keyboard_customization file.

BUG=b:274443029
TEST=zmake build  rex -DCONFIG_INPUT=y -DCONFIG_CROS_EC_KEYBOARD_INPUT=y
TEST=recovery, retraining and init-off works reliably

Change-Id: Icfaaf9db11f4fd959dc76b4a19c2c0d72d4202cf