include: Rename GCC_VERSION to TOOLCHAIN_GCC_VERSION to match Zephyr
GCC_VERSION was defined by CrosEC only, so when using Zephyr
COMPILER_HAS_GENERIC_BUILTIN_OVERFLOW was not defined. As a result,
check_add_overflow() macro was implemented using
`__builtin_choose_expr()` which is not provided by GNU C++ compiler.

Fortunately, Zephyr provides TOOLCHAIN_GCC_VERSION which is the same as
GCC_VERSION from CrosEC. To fix the problem, we rename GCC_VERSION to
TOOLCHAIN_GCC_VERSION in the EC codebase and provide it only if not
available.

BUG=b:299924239
TEST=make BOARD=bloonchipper
TEST=zmake build bloonchipper
TEST=./twister -v -i -T zephyr/test/fingerprint/task --toolchain host
TEST=./twister -v -i -T zephyr/test/fingerprint/task --toolchain llvm
TEST=Run Gitlab CI locally

Change-Id: Ic2b344d0e2ec1ac2018f359b3d8931093762fd1f