shim: add a __popcountsi2 implementation when using no_libgcc
We currently build without libgcc on ARM, which causes some libgcc
builtin to be missing. __popcountsi2 is used to implement POPCOUNT(),
add a standalone implementation built when libgcc is missing so that
POPCOUNT() references link properly.

BUG=none
TEST=ran some test using POPCOUNT, it works
LOW_COVERAGE_REASON=b:321660668 this only builds for ARM, can't test it
under POSIX

Change-Id: I81662e6ae3ad9d940445030373f25aed433fbaf4