servo_v4p1: Add poweron config feature
This feature allows to set up servo ports on start up as user needs.
Implementation is based on similar features already used in servo FW:
serialname and MAC address in non volatile flash.
New configs are added to config_allowed list, as this feature is limited
to servo and there is no plans of migrating servo to Zephyr.

BUG=b:316402626 b:248386313
TEST=manually check if all paths works as expected, use Omnigul DUT

Change-Id: Ic525d4f3ed7565d48666977c29afc9e6e54af9e6