usbc: disconnect mux in hard reset flow
In rex/screebo it was discovered that if a type-c port experiences
a hard reset while in dp alt mode with a titan ridge tbt3 dock that
the hard reset will not work properly. Everything appears to be
connected properly after the reset but the Aux lines do not work
properly. From captures it was determined that there is no mux/retimer
disconnect in the flow when this occurs and adding this seems to fix
the issues.

BUG=b:315354054
TEST=zmake rex/screebo verify that hard reset is no longer causing
the displays to not come up on resume.

Change-Id: If51f4bbe3f0ffcd6aec02063b9f3eaab318b31b6