sensors: Implement runtime interrupt switching
Add configuration value for a runtime swap between interrupt and forced
modes. Update motion_sense to leverage new driver API to change the mode
at runtime.

Example: if the threshold is set to 50ms any driver with the API function
implemented that's running faster than 20Hz will run in interrupt mode.
If the sensor is running at or slower than 20Hz, it will run in force
mode. If the sensor is explicitly listed in the force mode mask, it will
always run in force mode.

BUG=b:319290529,b:319290929
TEST=zmake build rex rex-ish-ec rex-ish brox brox-ish-ec
TEST=test_that $DUT_IP cheets_CTS_T.internal.x86.CtsSensor
LOW_COVERAGE_REASON=Coverage is added in chromium:5273875

Change-Id: I56aa932236b58d53c082b6ac1ce872db89ed587b