Prevent RW_FWID addr from ever changing
To prevent accidental changes in RW_FWID addresses, enforce that:

1) If there is no RW_FWID assertion in the BUILD.py, fail the build.
2) If there is a RW_FWID assertion, check the address, and fail the
build if it changes.

This only applies to boards that use binman. Other single image builds
and tests are not affected.

This could be generally useful for any other items where they must not
change after the first RO release.

BUG=b:269519200
TEST=zmake build -a with and without assertions, and with wrong numbers

Change-Id: I5da2105a3f067e8c209450b7b7b45ce2a085a154