ite: fix the location of FMAP
Pin the location of FMAP on ITE based projects to exist just before the
pstate section. This speeds up flash_ec script when sections must be
preserved.

The brox change must be part of this CL so that the default FMAP section
doesn't overlap with the brox pstate location.

Brox FMAP:
  $ dump_fmap -F build/zephyr/brox/output/ec.bin
  0x00000000:0x0005ffff WP_RO
  0x00000000:0x0005de05 EC_RO
  0x00000000:0x00032f63 RO_FW
  0x00032f64:0x00032f83 RO_FRID
  0x0005dc00:0x0005de05 FMAP
  0x0005e000:0x0005efff PSTATE
  0x0005f000:0x0005ffff CBI
  0x00060000:0x000bffff EC_RW
  0x00060000:0x000bffff RW_FW
  0x00060000:0x00092f63 ZEPHYR_RW
  0x000bffe0:0x000bffff RW_FWID

BUG=b:317306098
TEST=flash brox image

Change-Id: Id0cb526b41f060437b0274a159ef5bb346cfd925