anraggar: Modify the keyboard scan time
When we use shortcut keys to switch to display mode, EC
needs to handle the dock's pd_attention and keyboard scan
events at the same time. This results in EC watchdog.
Considering the degree of impact, modifying the keyboard
scan time has the least impact. Therefore, change the
keyboard scan time to 2ms.

BUG=b:322731050
TEST=Switching display mode, EC isn't reset.

Change-Id: I58fd2f0a2bc85d523102c7494aa35615ced8241a