brox: Add EC_CMD_GET_PD_PORT_CAPS implementation
We can re-use the existing implementation of the host command. Move it
to a new `usb_pd_host_cmd_common.c` file and combine in the
recently-created `pd_host_event.c` to keep related code together.

This CL doesn't touch any of the Kconfig flags used to control the
capabilities; these should be adjusted as features actually become
available.

BUG=b:322835380
TEST=zmake build brox brya; BOARD=volteer make;
TEST=./twister -s drivers/drivers.host_cmd
TEST=ectool on DUT

Change-Id: I9678ea5e45405b2fb38eded93152dc917e70387d