zephyr: fpmcu: update HOOT_TICK interval
Change the HOOK_TICK interval from 250ms to 500ms. The HOOK_TICK is used
only to reload the watchdog, which has 3s warning and 10s deadline.

This change limits amount of wakeups in the sleep mode, which reduces
power consumption.

Due to lack of continuous power consumption measurement, it is hard to
measure the actual reducement.

BUG=b:320622825
TEST=Make sure power consumption is slightly reduced in the sleep mode

Change-Id: I0364486694158147af143cd460f2cefc2ebc97fe