flash_ec: Fix FMAP section detection
When reading the FMAP section out of the currently running EC image, the
script was using "dump_fmap" to verify it was valid.  However,
"dump_fmap" succeeds as long as there is an FMAP header anywhere in the
file, even if the extracted FMAP section is incomplete.

Switch to checking that FMAP header is found at the beginning of the
extracted file.

BUG=none
TEST=./util/flash_ec --brox --zephyr

Change-Id: I934b0612008ffe09f35bd90e5d8885a8b7dc74ff