util: Catch specific exception in build_with_clang.py
The purpose of this is to pacify the preupload hook linter.

Only catch build failures which manifest as subprocess.run seeing a
non-zero exit status. It throws subprocess.CalledProcessError when
run is called with check=True.

We will crash the program and report the exception to console if any
other odd exception happens.

https://docs.python.org/3/library/subprocess.html

BUG=b:176500425
TEST=I changed the cmd variable to cmd=["false"] in the build function
     and checked that board were added to the failed_boards variable.
TEST=./util/build_with_clang.py

Change-Id: I3193e6562cce9def3db5c8cd1b373d4cceb5619f