pdc_power_mgmt: Check return code in `pdc info`
The return code from pdc_power_mgmt_get_info() is ignored in the `pdc
info` console command, so the output can display garbage data if the
underlying command had an issue.

Check the return code and display an error message if it is non-zero.

BUG=None
TEST=zmake build brox
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Id892b6e2fe730f4b10a2329c95d3972935043e83