PPM: Add UCSI Platform Policy Manager
This CL adds a UCSI PPM. It receives commands from the OPM and sends
them to a PDC through a PDC driver (e.g. pdc_rts54xx.c).

The PDM (i.e. pdc_power_mgmt.c) will continue to work like before. It
will get -EBUSY if a PDC driver is serving for another request.  There
is no change in this behavior but contention will be more frequent
because of the PPM.

BUG=b:319124515
TEST=Boot modified kernel on Brox. Observed UCSI commands from OPM were
     successfully processed.
LOW_COVERAGE_REASON=b:335261254

Change-Id: I7b01e2604f6e33125bd784854fc17ada581f6869