board/hyperdebug: Bugfix to custom JTAG pins
A console command allows changing which pins HyperDebug use for JTAG.
In case openocd or similar tool has left the JTAG connection open, this
CL adds code to close it before reconfiguring pins.  This is necessary
to ensure that the new pins are properly configured the next time
openocd opens a JTAG connection.

BUG=none
TEST=make build BOARD=hyperdebug

Change-Id: I956bf5ee1aa80845af301556352d7fdd6cb2fa21