usbpd: Move EC_CMD_TYPEC_CONTROL to common file, add stubs for PDC case
EC_CMD_TYPEC_CONTROL is needed for TCPMv2 and PDC-based devices. Move
the implementation for this HC to the common file and update Kconfig and
legacy config to keep it off in case of TCPMv1. Also update legacy tests
and fuzzing builds.

Add stubs for select functions in the PDC use case for unsupported
functionality.

BUG=b:325079650
TEST=build Brox, volteer (legacy TCPMv2), hatch (legacy TCPMv1)

Change-Id: I5b6a4d44895336104c01eb587f61a33a27446db6