zephyr: add support for tests on hardware
Add possibility to include tests in an image built with zmake.

The tests can run via shell command or are run automatically depending
on a chosen framework and configs.

The tests can be included in RO and RW or only RW copy.

BUG=b:324574796
TEST=Enable hardware tests with CONFIG_HW_TEST, shell support with
CONFIG_ZTEST_SHELL, Host Commands on the main thread
CONFIG_TASK_HOSTCMD_THREAD_MAIN and run test with the 'ztest run-all'
shell command

Change-Id: I53255e272e1018c91383bca42d7dddba7a429f5e