Makefile.rules: Add full RW/RO elf disassembly support for C++
This disassembly output is for humans to use when debugging.
For this reason, we should enable the source interleaving to make
it easier to identify what code causes which instructions.

Additionally, since we have C++ for some builds we need to enable the
demangling option.

These two options does add a bit of processing time, but they are outside
the normal build flow.

BUG=b:176500425
TEST=make BOARD=bloonchipper dis -j
TEST=make BOARD=dartmonkey dis -j
TEST=make BOARD=hatch dis -j

Change-Id: If7c84211c19d0a727f226610d87d0e6f838511df