UCSI/PPM: Update GET_PD_MESSAGE to GET_VDO command mapping
Update GET_PD_MESSAGE to GET_VDO command mapping to correctly route the
GET_VDO command to SOP, SOP' or SOP'' based on the recipient field of
the GET_PD_MESSAGE command. Additionally, request a placeholder VDO to
account for GET_PD_MESSAGE returning the VDM header with the discover
identity response. The VDM header is not used in chromeOS.

BUG=b:319506149
TEST=make um_ppm, confirmed the partner and cable identities were
correctly registered with the Linux Type-C connector class with a
modified v6.6 kernel image.

Change-Id: Idb2651afb9a4107cf54672cb3792525fbb4a7e18