board/hyperdebug: Better recording of bitbanged pins
Sometimes tests may do GPIO bitbanging on pins for which GPIO monitoring
is enabled.  This could be in order to record reaction on other pins for
correlation with the stimulating pins, or it could be simply to leave a
trace of what happened for developers to look at after completed tests.

Until now, the bitbanging timer interrupt by default was assigned the
highest priority level, meaning that edge detection interrupts would be
delayed until after the bitbanging handler had finished modifying up to
7 pins.  The order at which the edge handlers would run would then
probably be determined based on their pin number, not the order at which
the edges actually occurred.

With this change, the bitbanging timer interrupt is reduced to second
highest priority, allowing any edge detection interrupts to run as the
bitbanging is looping over the set of pins to modify, this allows
accurately capturing the order of edge events.

BUG=b:286553244
TEST=make tast TAST_EXPR=gscdevboard.GSCTPMI2CCorners

Change-Id: I62920a5ca78d69fccba2ead88faaf45f5cb2e86d