ec_features: Enable RWSIG feature if CONFIG_PLATFORM_EC_RWSIG is set
To work properly, RWSIG feature bit must be reported by RO and RW part
of the firmware. For RO part the bit is set because HAS_TASK_RWSIG is
set, but for RW part there is no RWSIG task. In EC the problem was
solved by checking HAS_TASK_RWSIG_RO which is set when RO has this task
enabled. In Zephyr, we don't define HAS_TASK_RWSIG_RO. We should use
CONFIG_PLATFORM_EC_RWSIG which is enabled for both RO and RW. The option
enables CONFIG_PLATFORM_EC_VERIFY for RO and for RW it's a no-op.

BUG=b:323890397
TEST=Build firmware with 'zmake build bloonchipper' and flash it.
     Jump to RW and make sure that bit 30 reported by 'feat' command
     is set.

Change-Id: I829cc4a8f7798d6b4b34f3a41aea60dad4d41a0e